requirements.txt
bidict==0.23.1
cryptography==44.0.2
dash==3.0.4
dash_bootstrap_components==2.0.2
Flask==3.1.0
kiteconnect==5.0.1
pandas==2.2.3
pyotp==2.9.0
python-dotenv==1.1.0
Requests==2.32.3
selenium==4.31.0
SQLAlchemy==2.0.40
sqlalchemy_utils==0.41.2
twilio==9.5.2
webdriver_manager==4.0.2
psycopg2-binary
asyncpg
numpy~=2.2.5